+++ CHANGELOG.md
@@ -5,5 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- Mock tests
+- `Cluster` interface
+- Dependency injection in `cluster.New()` to allow passing ec2 & autoscaling sdk objects from outside.
+- `cluster.NodeIdsToAwsStrings()` to convert a list of `cluster.NodeID` objects to list of `aws.String` objects
+
 ### Fixed
 - Upscale if the number of running agents is less than the minimum count set by user.